requirements.txt
fastapi~=0.111.1
httpx~=0.27.0
redis~=5.0.7
starlette~=0.37.2
asyncio~=3.4.3
bson~=0.5.10
python-dotenv~=1.0.1
motor~=3.5.1
pymongo~=4.8.0
aiologger~=0.7.0
aiofiles~=24.1.0
pydantic~=2.9.2